+++ components/layout/accordion/Accordion.mjs
@@ -8,6 +8,7 @@ export default class Accordion extends AbstractComponent {
      * Creates a new component
      * @param {object} props Component properties
      * @param {boolean} props.opened
+     * @param {boolean} props.borderless
      * @param {string[]|HTMLElement[]|AbstractElement[]} children Component children
      * @param {string} theme Component color theme
      * @param {string} id Component ID
@@ -18,12 +19,14 @@ export default class Accordion extends AbstractComponent {
     }
 
     static _defaultProps = {
-        opened: false
+        opened: false,
+        borderless: false
     }
 
     static validateProps(props) {
         return (
-            typeof props.opened === 'boolean'
+            typeof props.opened === 'boolean' &&
+            typeof props.borderless === "boolean"
         )
     }
 
@@ -31,6 +34,7 @@ export default class Accordion extends AbstractComponent {
 
     static _rawStylingRules = [
         `.accordion{--accordion-bg:var(--bg);--accordion-border:var(--border);box-shadow:var(--box-shadow);background-color:var(--accordion-bg);border:var(--border-thin-width) solid var(--accordion-border);border-radius:0;padding:var(--padding-sm);margin:var(--padding-sm)}`,
+        `.accordion.borderless{border:none;box-shadow:none}`,
         `.accordion.kw-{{theme}} .accordion-mark{color:var(--{{theme}})}`,
         `.accordion>:not(.accordion-sum){transition: max-height var(--transition-slow) ease-in-out;max-height:1000px;overflow:hidden}`,
         `.accordion:not([open])>:not(.accordion-sum){max-height:0;transition:none}`
@@ -40,7 +44,16 @@ export default class Accordion extends AbstractComponent {
         if (this.isMounted() && JSON.stringify(this._props) !== JSON.stringify(oldProps)) {
             let i = this.i();
 
-            (oldProps.opened !== this.props.opened) && this.props.opened ? i.setAttribute("open", "") : i.removeAttribute("open");
+            if (oldProps.opened !== this.props.opened) {
+                i.removeAttribute("open");
+
+                if (this.props.opened) i.setAttribute("open", "");
+            }
+            if (oldProps.borderless !== this.props.borderless) {
+                i.classList.remove("borderless");
+
+                if (this.props.borderless) i.classList.add("borderless");
+            }
         };
     }
 
@@ -50,6 +63,7 @@ export default class Accordion extends AbstractComponent {
         el.classList.add("accordion", this.themeClass());
 
         if (this.props.opened) el.setAttribute("open", "");
+        if (this.props.borderless) el.classList.add("borderless");
 
         this.attachChildren(el);
         this.attachListeners(el);